Refactor MentionTile customisation options
