+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -4,14 +4,14 @@
 
 - `StreamMessageComposer` now surfaces the hold-to-record hint through `StreamSnackbar` anchored above the composer, and `StreamChat` provides an app-wide `StreamSnackbarScope` fallback.
 - Added support for `@channel`, `@here`, role, and user-group mentions — parsed on incoming messages, rendered as styled tappable spans in message text, and selectable from the composer's `@` autocomplete.
-- Added per-type tile builders on `StreamMessageComposer` and `StreamMentionAutocompleteOptions` (`mentionChannelTileBuilder`, `mentionHereTileBuilder`, `mentionRoleTileBuilder`, `mentionUserGroupTileBuilder`, `mentionUserTileBuilder`), plus matching `onMention*Tap` callbacks on `StreamMentionAutocompleteOptions`.
+- Added a single `mentionTileBuilder` on `StreamMessageComposer` and `StreamMentionAutocompleteOptions` that receives `StreamMentionTileProps` and covers every mention kind. Customise globally via `streamChatComponentBuilders(mentionTile: ...)` or per-instance via the new constructor parameter. Defaults are rendered by `DefaultStreamMentionTile`. Also added `onMention*Tap` callbacks on `StreamMentionAutocompleteOptions`.
 - Added `StreamMessageListView.onMentionTap` and `StreamMessageItem.onMentionTap` — receives a typed `StreamMention` (`StreamUserMention`, `StreamChannelMention`, `StreamHereMention`, `StreamRoleMention`, or `StreamGroupMention`).
 
 ⚠️ Deprecated
 
 - `StreamAudioRecorderController.showInfo` is now deprecated. Show your own snackbar via `StreamSnackbarMessenger.of(context).show(StreamSnackbar(...))` instead.
 - `RecordStateIdle.message` is now deprecated; the composer no longer reads it.
-- Deprecated `userMentionsTileBuilder` on `StreamMessageComposer` and `mentionsTileBuilder` on `StreamMentionAutocompleteOptions` in favor of `mentionUserTileBuilder`.
+- Deprecated `userMentionsTileBuilder` on `StreamMessageComposer` and `mentionsTileBuilder` on `StreamMentionAutocompleteOptions` in favor of `mentionTileBuilder`.
 - Deprecated `StreamMessageListView.onUserMentionTap` and `StreamMessageItem.onUserMentionTap` in favor of `onMentionTap`.
 - Deprecated `UserListX.search`.
