+++ packages/stream_chat_flutter/lib/src/autocomplete/stream_mention_autocomplete_options.dart
@@ -19,12 +19,8 @@ class StreamMentionAutocompleteOptions extends StatefulWidget {
     this.client,
     this.limit = 10,
     this.mentionAllAppUsers = false,
-    @Deprecated('Use mentionUserTileBuilder instead') this.mentionsTileBuilder,
-    this.mentionChannelTileBuilder,
-    this.mentionHereTileBuilder,
-    this.mentionRoleTileBuilder,
-    this.mentionUserGroupTileBuilder,
-    this.mentionUserTileBuilder,
+    @Deprecated('Use mentionTileBuilder instead') this.mentionsTileBuilder,
+    this.mentionTileBuilder,
     this.onMentionChannelTap,
     this.onMentionHereTap,
     this.onMentionRoleTap,
@@ -57,24 +53,21 @@ class StreamMentionAutocompleteOptions extends StatefulWidget {
   /// Defaults to false.
   final bool mentionAllAppUsers;
 
-  /// Customize the tile for the mentions overlay.
-  @Deprecated('Use mentionUserTileBuilder instead')
+  /// Customize the tile for user mentions in the overlay.
+  ///
+  /// Honoured only for user mentions, for backwards compatibility. Prefer
+  /// [mentionTileBuilder] (or a globally-registered factory builder), which
+  /// covers every mention kind.
+  @Deprecated('Use mentionTileBuilder instead')
   final UserMentionTileBuilder? mentionsTileBuilder;
 
-  /// Builder for a custom `@channel` broadcast mention tile.
-  final Widget Function(BuildContext context)? mentionChannelTileBuilder;
-
-  /// Builder for a custom `@here` broadcast mention tile.
-  final Widget Function(BuildContext context)? mentionHereTileBuilder;
-
-  /// Builder for a custom role mention tile.
-  final MentionRoleTileBuilder? mentionRoleTileBuilder;
-
-  /// Builder for a custom user group mention tile.
-  final MentionUserGroupTileBuilder? mentionUserGroupTileBuilder;
-
-  /// Builder for a custom user mention tile.
-  final UserMentionTileBuilder? mentionUserTileBuilder;
+  /// Per-instance builder for the suggestion tile rendered for each mention.
+  ///
+  /// Takes precedence over any builder registered globally for
+  /// [StreamMentionTileProps] through [streamChatComponentBuilders]. When
+  /// null, the global builder is consulted next, then [DefaultStreamMentionTile]
+  /// is used as a fallback.
+  final StreamMentionTileBuilder? mentionTileBuilder;
 
   /// Callback called when the "@channel" broadcast mention is selected.
   final VoidCallback? onMentionChannelTap;
@@ -139,10 +132,26 @@ class _StreamMentionAutocompleteOptionsState extends State<StreamMentionAutocomp
           margin: margin,
           shape: shape,
           optionBuilder: (context, option) => switch (option) {
-            _ChannelMention() => _buildChannelTile(context),
-            _HereMention() => _buildHereTile(context),
-            _RoleMention(role: final role) => _buildRoleTile(context, role),
-            _UserGroupMention(userGroup: final userGroup) => _buildUserGroupTile(context, userGroup),
+            _ChannelMention() => _buildTile(
+              context,
+              const StreamChannelMention(),
+              widget.onMentionChannelTap,
+            ),
+            _HereMention() => _buildTile(
+              context,
+              const StreamHereMention(),
+              widget.onMentionHereTap,
+            ),
+            _RoleMention(role: final role) => _buildTile(
+              context,
+              StreamRoleMention(role: role.name),
+              widget.onMentionRoleTap == null ? null : () => widget.onMentionRoleTap!(role),
+            ),
+            _UserGroupMention(userGroup: final userGroup) => _buildTile(
+              context,
+              StreamGroupMention(userGroup: userGroup),
+              widget.onMentionUserGroupTap == null ? null : () => widget.onMentionUserGroupTap!(userGroup),
+            ),
             _UserMention(user: final user) => _buildUserTile(context, user),
           },
         );
@@ -150,76 +159,28 @@ class _StreamMentionAutocompleteOptionsState extends State<StreamMentionAutocomp
     );
   }
 
-  Widget _buildChannelTile(BuildContext context) {
-    final builder = widget.mentionChannelTileBuilder;
-    if (builder != null) {
-      return Material(
-        color: context.streamColorScheme.backgroundElevation1,
-        child: InkWell(
-          onTap: widget.onMentionChannelTap,
-          child: builder(context),
-        ),
-      );
-    }
-    return _MentionChannelTile(onTap: widget.onMentionChannelTap);
-  }
-
-  Widget _buildHereTile(BuildContext context) {
-    final builder = widget.mentionHereTileBuilder;
-    if (builder != null) {
-      return Material(
-        color: context.streamColorScheme.backgroundElevation1,
-        child: InkWell(
-          onTap: widget.onMentionHereTap,
-          child: builder(context),
-        ),
-      );
-    }
-    return _MentionHereTile(onTap: widget.onMentionHereTap);
-  }
-
-  Widget _buildRoleTile(BuildContext context, Role role) {
-    final tap = widget.onMentionRoleTap == null ? null : () => widget.onMentionRoleTap!(role);
-    final builder = widget.mentionRoleTileBuilder;
-    if (builder != null) {
-      return Material(
-        color: context.streamColorScheme.backgroundElevation1,
-        child: InkWell(
-          onTap: tap,
-          child: builder(context, role),
-        ),
-      );
-    }
-    return _MentionRoleTile(role: role.name, onTap: tap);
-  }
-
-  Widget _buildUserGroupTile(BuildContext context, UserGroup userGroup) {
-    final tap = widget.onMentionUserGroupTap == null ? null : () => widget.onMentionUserGroupTap!(userGroup);
-    final builder = widget.mentionUserGroupTileBuilder;
-    if (builder != null) {
-      return Material(
-        color: context.streamColorScheme.backgroundElevation1,
-        child: InkWell(
-          onTap: tap,
-          child: builder(context, userGroup),
-        ),
-      );
-    }
-    return _MentionUserGroupTile(userGroup: userGroup, onTap: widget.onMentionUserGroupTap);
+  Widget _buildTile(BuildContext context, StreamMention mention, VoidCallback? onTap) {
+    final props = StreamMentionTileProps(mention: mention, onTap: onTap);
+    final localBuilder = widget.mentionTileBuilder;
+    if (localBuilder != null) return localBuilder(context, props);
+    return StreamMentionTile.fromProps(props: props);
   }
 
   Widget _buildUserTile(BuildContext context, User user) {
-    final builder = widget.mentionUserTileBuilder ?? widget.mentionsTileBuilder;
-    if (builder != null) {
+    final tap = widget.onMentionUserTap == null ? null : () => widget.onMentionUserTap!(user);
+    // Deprecated legacy builder: honoured only for user mentions.
+    // ignore: deprecated_member_use_from_same_package
+    final legacy = widget.mentionsTileBuilder;
+    if (legacy != null && widget.mentionTileBuilder == null) {
       return Material(
         color: context.streamColorScheme.backgroundElevation1,
         child: InkWell(
-          onTap: widget.onMentionUserTap == null ? null : () => widget.onMentionUserTap!(user),
-          child: builder(context, user),
+          onTap: tap,
+          child: legacy(context, user),
         ),
       );
     }
-    return _MentionUserTile(user: user, onTap: widget.onMentionUserTap);
+    return _buildTile(context, StreamUserMention(user: user), tap);
   }
 
   List<User> get membersAndWatchers {
@@ -328,186 +289,6 @@ class _StreamMentionAutocompleteOptionsState extends State<StreamMentionAutocomp
       sort: [const SortOption.asc('id')],
     );
     return response.users;
-  }
-}
-
-/// Shared row layout for every mention autocomplete tile.
-class _MentionTile extends StatelessWidget {
-  const _MentionTile({
-    required this.leading,
-    required this.title,
-    this.subtitle,
-    this.onTap,
-  });
-
-  final Widget leading;
-  final String title;
-  final String? subtitle;
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final spacing = context.streamSpacing;
-    final colorScheme = context.streamColorScheme;
-    final textTheme = context.streamTextTheme;
-    final subtitle = this.subtitle;
-    return Padding(
-      padding: EdgeInsets.symmetric(horizontal: spacing.xxs),
-      child: InkWell(
-        borderRadius: BorderRadius.circular(8),
-        onTap: onTap,
-        child: ConstrainedBox(
-          constraints: const BoxConstraints(minHeight: 48),
-          child: Padding(
-            padding: EdgeInsets.symmetric(
-              horizontal: spacing.sm,
-              vertical: spacing.xs,
-            ),
-            child: Row(
-              spacing: spacing.sm,
-              children: [
-                leading,
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    spacing: spacing.xxxs,
-                    children: [
-                      Text(
-                        title,
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: textTheme.bodyDefault,
-                      ),
-                      if (subtitle != null)
-                        Text(
-                          subtitle,
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                          style: textTheme.metadataDefault.copyWith(
-                            color: colorScheme.textTertiary,
-                          ),
-                        ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Circular icon badge used as the leading widget for non-user mention tiles.
-class _MentionIconBadge extends StatelessWidget {
-  const _MentionIconBadge({required this.icon});
-
-  final IconData icon;
-
-  @override
-  Widget build(BuildContext context) {
-    final spacing = context.streamSpacing;
-    final colorScheme = context.streamColorScheme;
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: colorScheme.backgroundSurfaceSubtle,
-        border: Border.all(color: colorScheme.borderSubtle),
-        shape: BoxShape.circle,
-      ),
-      child: Padding(
-        padding: EdgeInsets.all(spacing.xs),
-        child: Icon(icon, size: 16),
-      ),
-    );
-  }
-}
-
-/// Represents the tile for "@channel" mention.
-class _MentionChannelTile extends StatelessWidget {
-  const _MentionChannelTile({this.onTap});
-
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return _MentionTile(
-      leading: _MentionIconBadge(icon: context.streamIcons.megaphone),
-      title: '@channel',
-      subtitle: context.translations.notifyChannelText,
-      onTap: onTap,
-    );
-  }
-}
-
-/// Represents the tile for "@here" mention.
-class _MentionHereTile extends StatelessWidget {
-  const _MentionHereTile({this.onTap});
-
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return _MentionTile(
-      leading: _MentionIconBadge(icon: context.streamIcons.megaphone),
-      title: '@here',
-      subtitle: context.translations.notifyHereText,
-      onTap: onTap,
-    );
-  }
-}
-
-/// Represents the tile for role mentions.
-class _MentionRoleTile extends StatelessWidget {
-  const _MentionRoleTile({required this.role, this.onTap});
-
-  final String role;
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return _MentionTile(
-      leading: _MentionIconBadge(icon: context.streamIcons.shield),
-      title: '@$role',
-      subtitle: context.translations.notifyRoleText(role),
-      onTap: onTap,
-    );
-  }
-}
-
-/// Represents the tile for user group mentions.
-class _MentionUserGroupTile extends StatelessWidget {
-  const _MentionUserGroupTile({required this.userGroup, this.onTap});
-
-  final UserGroup userGroup;
-  final ValueSetter<UserGroup>? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final description = userGroup.description;
-    return _MentionTile(
-      leading: _MentionIconBadge(icon: context.streamIcons.users),
-      title: '@${userGroup.name}',
-      subtitle: description?.isNotEmpty == true ? description : null,
-      onTap: onTap == null ? null : () => onTap!(userGroup),
-    );
-  }
-}
-
-/// Represents the tile for user mentions.
-class _MentionUserTile extends StatelessWidget {
-  const _MentionUserTile({required this.user, this.onTap});
-
-  final User user;
-  final ValueSetter<User>? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return _MentionTile(
-      leading: StreamUserAvatar(size: .md, user: user),
-      title: user.name,
-      onTap: onTap == null ? null : () => onTap!(user),
-    );
   }
 }
